+++ src/components/routes/home/Event.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const Event = () => {
+  return (
+    <div className="home-body-event">
+      <h3>이벤트</h3>
+    </div>
+  );
+};
+
+export default React.memo(Event);
