+++ src/components/routes/home/GreetAndPreScreen.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const GreetAndPreScreen = () => {
+  return (
+    <div className="home-body-theater">
+      <h3>무대인사 & 시사</h3>
+    </div>
+  );
+};
+
+export default React.memo(GreetAndPreScreen);
