+++ src/components/routes/home/MovieFeed.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const MovieFeed = React.memo(() => {
+  return (
+    <div className="home-body-feed">
+      <h3>무비피드</h3>
+    </div>
+  );
+});
+
+export default MovieFeed;
