+++ src/components/routes/home/MoviePost.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const MoviePost = () => {
+  return (
+    <div className="home-body-post">
+      <h3>무비 포스트</h3>
+    </div>
+  );
+};
+
+export default React.memo(MoviePost);
