+++ src/components/routes/home/Special.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const Special = () => {
+  return (
+    <div className="home-body-special">
+      <h3>특별관</h3>
+    </div>
+  );
+};
+
+export default React.memo(Special);
